+++ openframe/libs/openframe-data/src/main/java/com/openframe/data/config/KafkaConfig.java
@@ -1,11 +1,8 @@
 package com.openframe.data.config;
 
-import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.Map;
 
-import com.openframe.data.model.debezium.DeserializedDebeziumMessage;
-import com.openframe.data.model.enums.MessageType;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.common.serialization.StringDeserializer;
@@ -14,7 +11,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.convert.converter.Converter;
 import org.springframework.kafka.annotation.EnableKafka;
 import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
 import org.springframework.kafka.core.ConsumerFactory;
@@ -22,7 +18,6 @@ import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
 import org.springframework.kafka.core.DefaultKafkaProducerFactory;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.kafka.core.ProducerFactory;
-import org.springframework.kafka.support.converter.JsonMessageConverter;
 import org.springframework.kafka.support.serializer.JsonDeserializer;
 import org.springframework.kafka.support.serializer.JsonSerializer;
 
@@ -83,53 +78,5 @@ public class KafkaConfig {
                 new ConcurrentKafkaListenerContainerFactory<>();
         factory.setConsumerFactory(consumerFactory());
         return factory;
-    }
-
-    @Bean
-    public ConcurrentKafkaListenerContainerFactory<String, DeserializedDebeziumMessage> debeziumKafkaListenerContainerFactory() {
-        ConcurrentKafkaListenerContainerFactory<String, DeserializedDebeziumMessage> factory =
-                new ConcurrentKafkaListenerContainerFactory<>();
-
-        Map<String, Object> props = new HashMap<>();
-        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-        props.put(ConsumerConfig.GROUP_ID_CONFIG, "event-processor");
-        props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
-        props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, JsonDeserializer.class);
-        props.put(ConsumerConfig.MAX_PARTITION_FETCH_BYTES_CONFIG, 10485760);
-        props.put(ConsumerConfig.FETCH_MAX_BYTES_CONFIG, 10485760);
-        props.put(ConsumerConfig.RECEIVE_BUFFER_CONFIG, 10485760);
-        props.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
-        props.put(JsonDeserializer.TRUSTED_PACKAGES, "*");
-        props.put(JsonDeserializer.USE_TYPE_INFO_HEADERS, false);
-        props.put(JsonDeserializer.VALUE_DEFAULT_TYPE, "com.openframe.data.model.debezium.DeserializedDebeziumMessage");
-        props.put(JsonDeserializer.TYPE_MAPPINGS, "debeziumMessage:com.openframe.data.model.debezium.DeserializedDebeziumMessage");
-
-        ConsumerFactory<String, DeserializedDebeziumMessage> consumerFactory =
-                new DefaultKafkaConsumerFactory<>(props, new StringDeserializer(),
-                        new JsonDeserializer<>(DeserializedDebeziumMessage.class));
-
-        factory.setConsumerFactory(consumerFactory);
-        return factory;
-    }
-
-    @Bean
-    public JsonMessageConverter jsonMessageConverter() {
-        return new JsonMessageConverter();
-    }
-
-    @Bean
-    public Converter<byte[], MessageType> messageTypeConverter() {
-        return new Converter<byte[], MessageType>() {
-            @Override
-            public MessageType convert(byte[] source) {
-                try {
-                    String stringValue = new String(source, StandardCharsets.UTF_8);
-                    return MessageType.valueOf(stringValue.toUpperCase());
-                } catch (IllegalArgumentException e) {
-                    return null;
-                }
-            }
-
-        };
     }
 }
