+++ openframe/libs/openframe-data/src/main/java/com/openframe/data/model/debezium/DebeziumMessage.java
@@ -1,23 +1,26 @@
 package com.openframe.data.model.debezium;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.JsonNode;
 import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
 
 
 @Data
-public class DebeziumMessage {
+@SuperBuilder
+@NoArgsConstructor
+public class DebeziumMessage<T> {
 
     @JsonProperty("payload")
-    private Payload payload;
+    private Payload<T> payload;
 
     @Data
-    public static class Payload {
+    public static class Payload<T> {
         @JsonProperty("before")
-        private JsonNode before;
+        private T before;
 
         @JsonProperty("after")
-        private JsonNode after;
+        private T after;
 
         @JsonProperty("source")
         private Source source;
