+++ openframe/libs/openframe-data/src/main/java/com/openframe/data/model/debezium/DeserializedDebeziumMessage.java
@@ -3,11 +3,13 @@ package com.openframe.data.model.debezium;
 import com.openframe.data.model.enums.IntegratedToolType;
 import com.openframe.data.model.enums.UnifiedEventType;
 import lombok.Data;
+import lombok.experimental.SuperBuilder;
 
 import java.util.Map;
 
 @Data
-public class DeserializedDebeziumMessage extends DebeziumMessage {
+@SuperBuilder
+public class DeserializedDebeziumMessage extends CommonDebeziumMessage {
 
     private UnifiedEventType unifiedEventType;
     private String ingestDay;
