+++ openframe/services/openframe-stream/src/main/java/com/openframe/stream/deserializer/KafkaMessageDeserializer.java
@@ -1,5 +1,6 @@
 package com.openframe.stream.deserializer;
 
+import com.openframe.data.model.debezium.CommonDebeziumMessage;
 import com.openframe.data.model.debezium.DeserializedDebeziumMessage;
 import com.openframe.data.model.enums.MessageType;
 
@@ -7,6 +8,6 @@ public interface KafkaMessageDeserializer {
 
     MessageType getType();
 
-    DeserializedDebeziumMessage deserialize(DeserializedDebeziumMessage message, MessageType type);
+    DeserializedDebeziumMessage deserialize(CommonDebeziumMessage message, MessageType type);
 
 }
